Added Accumulated agents visualization using plane grid

+++ Swarm/Assets/Scripts/SwarmSimulator/Grid/Components/GridDotTag.cs
@@ -4,6 +4,6 @@ namespace Swarm.Grid
 {
     public struct GridDotTag : IComponentData
     {
-        public float Index;
+        public int Index;
     }
 }
